Send 'OK' back to PC when LI is addresses again.
Postpone sending of an 'OK' response from a moment when command was sent to
a command station to the moment when the command station sends 'Normal
Inquiry' after an outgoing command was sent.

This allows the command station to respond with a message (possibly error).
When the command station answer with a message for us, 'OK' response
is not send (but the message is).

This commit removes inspection of the PC->XN message and determining whether to
respond 'OK' or not.